mostly finish all classes
